requirements.txt
flask
python-telegram-bot
Werkzeug
gunicorn
gevent
requests